project/scraping_project/requirements.txt
beautifulsoup4==4.12.3
colorthief==0.2.1
Django==5.1.3
matplotlib==3.9.2
numpy==2.1.3
pandas==2.2.3
Pillow==11.0.0
Requests==2.32.3
selenium==4.26.1
webdriver_manager==4.0.2
